+++ src/typechecker/resolver_validation.rs
@@ -13,6 +13,7 @@ include!("resolver_validation/imports_modules.rs");
 include!("resolver_validation/imports_dependencies.rs");
 include!("resolver_validation/imports_behavior_dependencies.rs");
 include!("resolver_validation/imports_source_dependencies.rs");
+include!("resolver_validation/imports_source_dependency_callables.rs");
 include!("resolver_validation/imported_method_seeding.rs");
 include!("resolver_validation/symbols_locals.rs");
 include!("resolver_validation/local_scope_helpers.rs");
