+++ tests/docs_truth/repo_hygiene/file_size.rs
@@ -6,5 +6,6 @@ mod generic_behavior_splits;
 mod lexer_and_monomorphize;
 mod module_system_splits;
 mod resolver_phase2_splits;
+mod resolver_validation_splits;
 mod thresholds;
 mod typechecker_test_splits;
